requirements.txt
langchain
langgraph
tavily-api
faiss-cpu
chromadb
openai
python-dotenv
